单机消息队列私人聊天室
